move parameter test into client
